+++ src/main/java/com/iwayee/activity/api/comp/Player.java
@@ -4,7 +4,7 @@ import io.vertx.core.json.JsonObject;
 
 // 参与者
 public class Player {
-  public int id;
+  public long id;
   public int sex;
   public String wx_nick;
   public String nick;
@@ -26,7 +26,7 @@ public class Player {
   }
 
   public void fromJson(JsonObject data) {
-    id = data.getInteger("id");
+    id = data.getLong("id");
     wx_nick = data.getString("wx_nick");
     nick = data.getString("nick");
     sex = data.getInteger("sex");
