+++ src/main/java/com/iwayee/activity/api/comp/Session.java
@@ -3,6 +3,6 @@ package com.iwayee.activity.api.comp;
 final public class Session {
   public long at; // 上次登录时间，毫秒数
   public int sex;
-  public int uid; // 用户ID
+  public long uid; // 用户ID
   public String token;
 }
